+++ samples/Generic.java
@@ -1,9 +1,20 @@
 package samples;
 
+import java.util.Collection;
+import java.util.List;
 import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.Set;
+import java.util.Map;
 
 public class Generic {
 
+    public Collection<Long> longs;
+    public List<Double> doubles;
     public ArrayList<String> strings;
+    public Iterator<Float> floats;
+    public Set<Boolean> booleans;
+
+    //public Map<String, Boolean> stringsBooleans;
 
 }