+++ samples/Generic.ts
@@ -1,5 +1,13 @@
 export class Generic {
 
+    public longs: Array<number>;
+
+    public doubles: Array<number>;
+
     public strings: Array<string>;
+
+    public floats: Array<number>;
+
+    public booleans: Array<boolean>;
 
 }