+++ src/ast/Type.js
@@ -1,6 +1,6 @@
 const Util = require('./Util');
 
-const primitifsType = {
+const primitivesType = {
     'void': 'void',
 
     'int': 'number',
@@ -19,29 +19,36 @@ const primitifsType = {
     'java.time.LocalDate': 'string',
     'java.time.LocalDateTime': 'string',
 
-    'java.util.ArrayList': 'Array'
+    'java.util.Collection': 'Array',
+    'java.util.List': 'Array',
+    'java.util.ArrayList': 'Array',
+    'java.util.Iterator': 'Array',
+    'java.util.Set': 'Array',
+
+    'java.util.Map': 'Map',
+    'java.util.HashMap': 'Map'
 };
 
 const regexpComposed = /([^<]+)<([^>]+)>/;
 
 const find = (javaType) => {
     console.log('Type', javaType);
-    if(primitifsType[javaType]) {
-        return {needImport: false, name: primitifsType[javaType]};
+    if(primitivesType[javaType]) {
+        return {needImport: false, name: primitivesType[javaType]};
     } else {
         const match = regexpComposed.exec(javaType);
         if(match) {
             const first = find(match[1]);
-            const second = find(match[2]);
-            const needImport = [first.needImport, second.needImport].reduce((acc, current) => {
-                if(current !== false && acc === false) {
-                    acc = current;
-                } else if(current !== false) {
-                    acc += current;
+            const seconds = match[2].split(',').map(f => find(f.trim()));
+            const needImport = seconds.concat(first).reduce((acc, current) => {
+                if(current.needImport !== false && acc === false) {
+                    acc = current.needImport;
+                } else if(current.needImport !== false) {
+                    acc += current.needImport;
                 }
                 return acc;
             }, false);
-            return {needImport: needImport, name: `${first.name}<${second.name}>`};
+            return {needImport: needImport, name: `${first.name}<${seconds.map(c => c.name).join(', ')}>`};
         } else if(!javaType.startsWith('java')) {
             const name = Util.nameClassWithoutPackage(javaType);
             return {needImport: `import {${name}} from './${name}';`, name: name};